+++ fabric21/src/main/java/com/vecoo/extraquests/ExtraQuests.java
@@ -23,7 +23,7 @@ public class ExtraQuests implements ModInitializer {
     private static ExtraQuests instance;
 
     private ServerConfig config;
-    private LocaleConfig locale;
+    private LocaleConfig localeConfig;
 
     private MinecraftServer server;
 
@@ -42,8 +42,8 @@ public class ExtraQuests implements ModInitializer {
         try {
             this.config = new ServerConfig();
             this.config.init();
-            this.locale = new LocaleConfig();
-            this.locale.init();
+            this.localeConfig = new LocaleConfig();
+            this.localeConfig.init();
         } catch (Exception e) {
             LOGGER.error("Error load config.", e);
         }
@@ -54,23 +54,23 @@ public class ExtraQuests implements ModInitializer {
         KeyValueReward.TYPE = RewardTypes.register(ResourceLocation.fromNamespaceAndPath(ExtraQuests.MOD_ID, "key_value"), KeyValueReward::new, () -> Icon.getIcon("minecraft:item/paper"));
     }
 
-    public static ExtraQuests getInstance() {
+    public static ExtraQuests instance() {
         return instance;
     }
 
-    public static Logger getLogger() {
+    public static Logger logger() {
         return LOGGER;
     }
 
-    public ServerConfig getConfig() {
+    public ServerConfig config() {
         return instance.config;
     }
 
-    public LocaleConfig getLocale() {
-        return instance.locale;
+    public LocaleConfig localeConfig() {
+        return instance.localeConfig;
     }
 
-    public MinecraftServer getServer() {
+    public MinecraftServer server() {
         return instance.server;
     }
 }