+++ forge20/src/main/java/com/vecoo/extraquests/task/KeyValueTask.java
@@ -62,11 +62,11 @@ public class KeyValueTask extends Task {
         this.value = buffer.readVarLong();
     }
 
-    public String getKey() {
+    public String key() {
         return this.key;
     }
 
-    public long getValue() {
+    public long value() {
         return this.value;
     }