+++ neoforge21/src/main/java/com/vecoo/extraquests/mixin/CommandRewardMixin.java
@@ -73,10 +73,10 @@ public abstract class CommandRewardMixin {
     )
     public void claim(Commands instance, CommandSourceStack source, String command) {
         if (this.console) {
-            ServerConfig config = ExtraQuests.getInstance().getConfig();
+            ServerConfig config = ExtraQuests.instance().config();
 
             if (config.isBlacklistConsole()) {
-                for (String blacklistCommand : config.getBlacklistConsoleList()) {
+                for (String blacklistCommand : config.blacklistConsoleList()) {
                     if (command.contains(blacklistCommand)) {
                         return;
                     }