+++ changelog.md
@@ -1,1 +1,3 @@
-* Fix crash on use command.
+* Added versioned config.
+* Change name methods.
+* Update depends.